move itemable to concerns
